[Refactor] Reorganized folder, update README, fix php
