--- requirements.txt
@@ -1,2 +1,0 @@
-requests==2.22.0
-tenacity==5.0.3
