+++ samples/py/constants.py
@@ -1,0 +1,12 @@
+KEY = "Replace With Your Musiio API KEY"
+
+BASE_URL = "Replace with Your API URL"
+
+TAG_URL = BASE_URL + "/api/v1/extract/tags"
+
+UPLOAD_URL = BASE_URL + "/api/v1/upload/file"
+
+VALID_TAGS = ["CONTENT TYPE", "QUALITY", "GENRE", "GENRE SECONDARY", "GENRE V3", "BPM", "BPM ALT", "BPM VARIATION", "KEY FLAT", "KEY FLAT SECONDARY", "KEY SHARP", "KEY SHARP SECONDARY", "MOOD", "MOOD SECONDARY",
+              "MOOD VALENCE", "ENERGY", "ENERGY VARIATION", "INSTRUMENTATION", "VOCAL PRESENCE", "VOCAL GENDER", "INSTRUMENT"]
+
+TAGS = ["CONTENT TYPE", "GENRE", "GENRE V3", "BPM", "KEY SHARP", "KEY FLAT", "MOOD", "ENERGY", "INSTRUMENTATION"]
