+++ samples/py/generate_tags.py
