+++ samples/py/requirements.txt
@@ -1,0 +1,2 @@
+requests==2.22.0
+tenacity==5.0.3
