+++ samples/py/tags_to_csv.py
